requirements.txt
asgiref==3.11.1
Django==6.0.2
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
graphene==3.4.3
graphene-django==3.2.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==25.1.0
packaging==26.0
promise==2.3
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.5
text-unidecode==1.3
typing_extensions==4.15.0
